Refactor notionBlockToMarkdown function
This commit wholelly refactors the notionBlockToMarkdown function.
Originally, my motivation for this was that I needed blockquotes
that could contain multiple paragraphs. However, I realized that my
approach fixed a few other problems, so I went ahead and did the whole
thing.

There are a few things this code does:

* The `reduce` function could be removed by relying on the recursive
  nature of the blocks.
* Ensures that most block elements are always separated by two newlines.
  This is important to prevent them from being interpreted as a single
  entry, such as with back-to-back blockquotes.
* Allows blockquotes to contain child blocks, such as lists and
  paragraphs.
* Allows list items to contain child blocks, such as lists and
  paragraphs.